+++ packages/@runlightyear/gsheets/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @runlightyear/gsheets
 
+## 0.4.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @runlightyear/lightyear@0.31.0
+
 ## 0.4.0
 
 ### Minor Changes