+++ packages/@runlightyear/notion/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @runlightyear/notion
 
+## 0.2.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @runlightyear/lightyear@0.31.0
+
 ## 0.2.0
 
 ### Minor Changes